Change TPC-C back to Decimal instead of Real. (#41)
Fix #39.

+++ src/main/java/com/oltpbenchmark/benchmarks/tpcc/procedures/Delivery.java
@@ -237,7 +237,7 @@ public class Delivery extends TPCCProcedure {
                 }
 
                 int idx = 1; // HACK: So that we can debug this query
-                delivUpdateCustBalDelivCnt.setDouble(idx++, ol_total);
+                delivUpdateCustBalDelivCnt.setBigDecimal(idx++, BigDecimal.valueOf(ol_total));
                 delivUpdateCustBalDelivCnt.setInt(idx++, w_id);
                 delivUpdateCustBalDelivCnt.setInt(idx++, d_id);
                 delivUpdateCustBalDelivCnt.setInt(idx, c_id);
